forgot ses name

+++ scripts/kill_current_session.sh
@@ -23,17 +23,26 @@ if [ "$1" != "$force_directive" ]; then
     echo "$(date) not called with force" >> /tmp/session_kill.log
     ses_count="$(tmux list-sessions | wc -l)"
     if [ "$ses_count" -eq 1 ]; then
-	continue=0
 	$CURRENT_DIR/kill_session_confirm.sh
 	exit 0
+	echo" $(date) ERROR: continued after confirm exit!" >> /tmp/session_kill.log
     fi
 fi
+
+ses_to_go="$(tmux display-message -p '#S')"
+
+if [ -z "$ses_to_go" ]; then
+   echo "$(date) ses_to_go empty" >> /tmp/session_kill.log
+   exit 0
+fi
+
+echo "$(date) ses_to_go [$ses_to_go]" >> /tmp/session_kill.log
 
 
 #  switch to random next session
 tmux switch-client -n &
 
-sleep 1
+sleep 2
 
 tmux kill-session -t "$ses_to_go"
